requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic-settings==2.10.1
streamlit==1.48.1
